Smaller and Faster Stack Objects 🎉
~ This commit updates the representation of Hinton Objects so that heavier objects are heap-allocated (and can be referenced using Rc and RefCell), decreasing the payload of any single Hinton Stack object from 168 bytes down to 40 bytes 🎉.

Before, even storing a single integer (i64) on the VM's stack would take 168 bytes of space, which was obviously wasteful. With this update, the size has been decreased to 40 bytes for objects (and references to objects) that live on that stack.

NOTE: Even though this update is great, because we are using Rc and RefCell to refer to those objects that live on the heap, this new model for object representation needs to be tested for cyclic references and memory leaks.

* Because instances are now stored on the heap and can be mutated using RefCell, this update also fixes an issue where assigning a property to an instance would not update the base instance, but only a cloned version.